+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Improved converter plugins to allow more flexible pilet embedded usage
 * Added runtime pilet metadata `basePath`
+* Added support for new debug utils (#397)
 
 ## 0.13.5 (August 17, 2021)